+++ index.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <link rel="stylesheet"  href="css/normalize.css">
   <link rel="stylesheet"  href="css/style.css">
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
@@ -229,8 +229,5 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script> 
-      console.log(generateOffers())
-  </script>
 </body>
 </html>
